Optimize code and refresh option.
